moved the random secret generation in pkg/pipeline/secrets
