+++ pkg/pipelines/secrets/generate_secret.go
@@ -1,4 +1,4 @@
-package secret
+package secrets
 
 import (
 	"crypto/rand"
