+++ pkg/pipelines/secrets/generate_secret_test.go
@@ -1,4 +1,4 @@
-package secret
+package secrets
 
 import (
 	"testing"
